Handle current time in seconds as long
This avoids the overflow of the integer in 2038. Old getXXX methods remain for now as they provide null-safe access to the values.

Closes #11053

+++ adapters/oidc/adapter-core/src/main/java/org/keycloak/adapters/AdapterDeploymentContext.java
@@ -364,17 +364,17 @@ public class AdapterDeploymentContext {
         }
 
         @Override
-        public void setNotBefore(int notBefore) {
+        public void setNotBefore(long notBefore) {
             delegate.setNotBefore(notBefore);
         }
 
         @Override
-        public int getNotBefore() {
+        public long getNotBefore() {
             return delegate.getNotBefore();
         }
 
         @Override
-        public void updateNotBefore(int notBefore) {
+        public void updateNotBefore(long notBefore) {
             delegate.setNotBefore(notBefore);
             getPublicKeyLocator().reset(this);
         }
