+++ core/src/main/java/org/keycloak/representations/oidc/OIDCClientRepresentation.java
@@ -86,7 +86,7 @@ public class OIDCClientRepresentation {
 
     private String request_object_encryption_enc;
 
-    private Integer default_max_age;
+    private Long default_max_age;
 
     private Boolean require_auth_time;
 
@@ -111,9 +111,9 @@ public class OIDCClientRepresentation {
     private String software_version;
 
     // OIDC Dynamic Client Registration Response
-    private Integer client_id_issued_at;
+    private Long client_id_issued_at;
 
-    private Integer client_secret_expires_at;
+    private Long client_secret_expires_at;
 
     private String registration_client_uri;
 
@@ -370,11 +370,11 @@ public class OIDCClientRepresentation {
         this.request_object_encryption_enc = request_object_encryption_enc;
     }
 
-    public Integer getDefaultMaxAge() {
+    public Long getDefaultMaxAge() {
         return default_max_age;
     }
 
-    public void setDefaultMaxAge(Integer default_max_age) {
+    public void setDefaultMaxAge(Long default_max_age) {
         this.default_max_age = default_max_age;
     }
 
@@ -434,19 +434,19 @@ public class OIDCClientRepresentation {
         this.software_version = softwareVersion;
     }
 
-    public Integer getClientIdIssuedAt() {
+    public Long getClientIdIssuedAt() {
         return client_id_issued_at;
     }
 
-    public void setClientIdIssuedAt(Integer clientIdIssuedAt) {
+    public void setClientIdIssuedAt(Long clientIdIssuedAt) {
         this.client_id_issued_at = clientIdIssuedAt;
     }
 
-    public Integer getClientSecretExpiresAt() {
+    public Long getClientSecretExpiresAt() {
         return client_secret_expires_at;
     }
 
-    public void setClientSecretExpiresAt(Integer client_secret_expires_at) {
+    public void setClientSecretExpiresAt(Long client_secret_expires_at) {
         this.client_secret_expires_at = client_secret_expires_at;
     }
