+++ model/infinispan/src/main/java/org/keycloak/models/sessions/infinispan/changes/sessions/CrossDCLastSessionRefreshListener.java
@@ -67,7 +67,7 @@ public class CrossDCLastSessionRefreshListener implements ClusterListener {
         lastSessionRefreshes.entrySet().stream().forEach((entry) -> {
             String sessionId = entry.getKey();
             String realmId = entry.getValue().getRealmId();
-            int lastSessionRefresh = entry.getValue().getLastSessionRefresh();
+            long lastSessionRefresh = entry.getValue().getLastSessionRefresh();
 
             // All nodes will receive the message. So ensure that each node updates just lastSessionRefreshes owned by him.
             if (shouldUpdateLocalCache(sessionId)) {
